backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
boto3==1.29.7
openai==1.3.7
pinecone-client==2.2.4
pypdf2==3.0.1
pandas==2.1.3
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
tiktoken==0.5.2

